# doc/api.yaml
...
openapi: 3.0.3
info:
    title: 'WASAText API'
    description: |-
        A modern chat application API that allows users to create groups,
        start conversations, send messages, react to messages, and manage
        contacts.

        Features include:
        - Real-time messaging with WebSocket support
        - Image messaging with automatic compression
        - Message reactions with toggle behavior (prevents duplicates)
        - Group chat management
        - Contact management
        - Message forwarding (including images)
        - User profiles with photos
    version: '1.2.0'
security:
    - bearerAuth: []

# Note: All authenticated endpoints (those requiring bearerAuth) will return 
# a 401 Unauthorized response if authentication fails. See the UnauthorizedError 
# response definition in components for details.
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: UUID
            description: |-
                Bearer token authentication using User UUID as the token.
                The token should be the UUID of an existing user in the system.
    schemas:
        User:
            type: object
            properties:
                id:
                    type: string
                    example: 'f2555a8a-2e66-4326-9588-20e7e298d615'
                    description: UUID of the user
                username:
                    type: string
                    example: 'Alice'
                    pattern: '^.*?$'
                    minLength: 3
                    maxLength: 16
                picture:
                    type: string
                    example: 'https://example.com/avatar.jpg'
                    description: URL to user's profile picture
            required:
                - id
                - username
        Conversation:
            type: object
            properties:
                id:
                    type: string
                    example: 'c0c0c0c0-0000-0000-0000-000000000001'
                    description: UUID of the conversation
                name:
                    type: string
                    example: 'Group Chat'
                    description: Name of the conversation (empty for 1:1 chats)
                picture:
                    type: string
                    example: 'https://example.com/group-avatar.jpg'
                    description: URL to conversation's picture
                participants:
                    type: array
                    items:
                        $ref: '#/components/schemas/User'
                    description: List of users in the conversation
            required:
                - id
                - participants
        Message:
            type: object
            properties:
                id:
                    type: string
                    example: '85f9cc23-48d7-448d-9172-7d5fd9cd4858'
                    description: UUID of the message
                senderId:
                    type: string
                    example: 'f2555a8a-2e66-4326-9588-20e7e298d615'
                    description: UUID of the message sender
                text:
                    type: string
                    example: 'Hello, how are you?'
                    description: Content of the message
                imageUrl:
                    type: string
                    example: 'data:image/jpeg;base64,/9j/4AAQSkZJRgABAQAAAQ...'
                    description: Base64 encoded image data with MIME type prefix (optional)
                senderUsername:
                    type: string
                    example: 'Alice'
                    description: Username of the message sender
                time:
                    type: string
                    example: '2025-01-01T12:00:00Z'
                    description: Timestamp when the message was sent
                reactions:
                    type: object
                    additionalProperties:
                        type: object
                        properties:
                            count:
                                type: integer
                                example: 3
                            users:
                                type: array
                                items:
                                    type: string
                                example: ['Alice', 'Bob']
                    description: >-
                        Reactions to the message (emoji -> reaction data)
            required:
                - id
                - senderId
                - senderUsername
        Reaction:
            type: object
            properties:
                emoji:
                    type: string
                    example: '👍'
                    description: The emoji reaction to toggle (add if not present, remove if already exists)
            required:
                - emoji
        Comment:
            type: object
            properties:
                id:
                    type: string
                    example: 'comment-uuid'
                    description: UUID of the comment
                messageId:
                    type: string
                    example: 'message-uuid'
                    description: UUID of the parent message
                senderId:
                    type: string
                    example: 'user-uuid'
                    description: UUID of the comment sender
                content:
                    type: string
                    example: 'Great point!'
                    description: Content of the comment
            required:
                - id
                - messageId
                - senderId
                - content
...
        SendMessageRequest:
            type: object
            properties:
                content:
                    type: string
                    example: 'Hello everyone!'
                    description: Content of the message to send
                imageUrl:
                    type: string
                    example: 'data:image/jpeg;base64,/9j/4AAQSkZJRgABAQAAAQ...'
                    description: Base64 encoded image data with MIME type prefix (optional)
            anyOf:
                - required: [content]
                - required: [imageUrl]
...
        ForwardMessageRequest:
            type: object
            properties:
                content:
                    type: string
                    example: 'target-conversation-uuid'
                    description: UUID of the target conversation
            required:
                - content
...
        CommentRequest:
            type: object
            properties:
                content:
                    type: string
                    example: 'Great message!'
                    description: Content of the comment
            required:
                - content
        Error:
            type: object
            properties:
                message:
                    type: string
                    example: 'Invalid input provided'
                    description: Error message describing what went wrong
            required:
                - message
...
paths:
...
    /users/{id}/conversations/{conversationId}/messages:
        parameters:
            - name: id
              in: path
              description: UUID of the user
              required: true
              schema:
                  type: string
                  format: uuid
            - name: conversationId
              in: path
              description: UUID of the conversation
              required: true
              schema:
                  type: string
                  format: uuid
        get:
            tags: ['Messages']
            summary: Get messages
            description: Get all messages from a conversation
            operationId: getMessages
            responses:
                '200':
                    description: List of messages
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/Message'
                '404':
                    description: Conversation not found
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
        post:
            tags: ['Messages']
            summary: Send message
            description: Send a new text message or image to the conversation. Images should be provided as base64-encoded data URLs.
            operationId: sendMessage
            requestBody:
                description: Message content (text and/or image)
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/SendMessageRequest'
                required: true
            responses:
                '201':
                    description: Message sent successfully
                    content:
                        application/json:
                            schema:
                                type: string
                                description: UUID of the created message
                '400':
                    description: Invalid message content
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                '404':
                    description: Conversation not found
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
    /users/{id}/conversations/{conversationId}/messages/{messageId}:
        parameters:
            - name: id
              in: path
              description: UUID of the user
              required: true
              schema:
                  type: string
                  format: uuid
            - name: conversationId
              in: path
              description: UUID of the conversation
              required: true
              schema:
                  type: string
                  format: uuid
            - name: messageId
              in: path
              description: UUID of the message
              required: true
              schema:
                  type: string
                  format: uuid
        delete:
            tags: ['Messages']
            summary: Delete message
            description: Delete a message (only by the sender)
            operationId: deleteMessage
            responses:
                '200':
                    description: Message deleted successfully
                '403':
                    description: Not authorized to delete this message
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                '404':
                    description: Message not found
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
    /users/{id}/conversations/{conversationId}/messages/{messageId}/forward:
        parameters:
            - name: id
              in: path
              description: UUID of the user
              required: true
              schema:
                  type: string
                  format: uuid
            - name: conversationId
              in: path
              description: UUID of the conversation
              required: true
              schema:
                  type: string
                  format: uuid
            - name: messageId
              in: path
              description: UUID of the message
              required: true
              schema:
                  type: string
                  format: uuid
        post:
            tags: ['Messages']
            summary: Forward message
            description: Forward a message (including images) to another conversation with a [Forwarded] prefix
            operationId: forwardMessage
            requestBody:
                description: Target conversation for forwarding
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ForwardMessageRequest'
                required: true
            responses:
                '200':
                    description: Message forwarded successfully
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Conversation'
                '400':
                    description: Invalid request data
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                '404':
                    description: Message or target conversation not found
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
    /users/{id}/conversations/{conversationId}/messages/{messageId}/reaction:
        parameters:
            - name: id
              in: path
              description: UUID of the user
              required: true
              schema:
                  type: string
                  format: uuid
            - name: conversationId
              in: path
              description: UUID of the conversation
              required: true
              schema:
                  type: string
                  format: uuid
            - name: messageId
              in: path
              description: UUID of the message
              required: true
              schema:
                  type: string
                  format: uuid
        post:
            tags: ['Reactions']
            summary: Toggle reaction
            description: Toggle a reaction on a message. If the user has already reacted with this emoji, the reaction will be removed. If not, it will be added.
            operationId: reactToMessage
            requestBody:
                description: Emoji to toggle
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Reaction'
                required: true
            responses:
                '200':
                    description: Reaction toggled successfully
                '400':
                    description: Invalid reaction data
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                '404':
                    description: Message not found
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
    /users/{id}/conversations/{conversationId}/messages/{messageId}/reaction/{emoji}:
        parameters:
            - name: id
              in: path
              description: UUID of the user
              required: true
              schema:
                  type: string
                  format: uuid
            - name: conversationId
              in: path
              description: UUID of the conversation
              required: true
              schema:
                  type: string
                  format: uuid
            - name: messageId
              in: path
              description: UUID of the message
              required: true
              schema:
                  type: string
                  format: uuid
            - name: emoji
              in: path
              description: The emoji reaction to remove
              required: true
              schema:
                  type: string
        delete:
            tags: ['Reactions']
            summary: Remove reaction
            description: Remove a specific reaction from a message
            operationId: removeReaction
            responses:
                '200':
                    description: Reaction removed successfully
                '404':
                    description: Message or reaction not found
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
    /users/{id}/conversations/{conversationId}/messages/{messageId}/comments:
        parameters:
            - name: id
              in: path
              description: UUID of the user
              required: true
              schema:
                  type: string
                  format: uuid
            - name: conversationId
              in: path
              description: UUID of the conversation
              required: true
              schema:
                  type: string
                  format: uuid
            - name: messageId
              in: path
              description: UUID of the message
              required: true
              schema:
                  type: string
                  format: uuid
        post:
            tags: ['Comments']
            summary: Add comment
            description: Add a comment to a message
            operationId: commentMessage
            requestBody:
                description: Comment content
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CommentRequest'
                required: true
            responses:
                '201':
                    description: Comment added successfully
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Comment'
                '400':
                    description: Invalid comment data
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                '404':
                    description: Message not found
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
    /users/{id}/conversations/{conversationId}/messages/{messageId}/comments/{commentId}:
        parameters:
            - name: id
              in: path
              description: UUID of the user
              required: true
              schema:
                  type: string
                  format: uuid
            - name: conversationId
              in: path
              description: UUID of the conversation
              required: true
              schema:
                  type: string
                  format: uuid
            - name: messageId
              in: path
              description: UUID of the message
              required: true
              schema:
                  type: string
                  format: uuid
            - name: commentId
              in: path
              description: UUID of the comment
              required: true
              schema:
                  type: string
                  format: uuid
        delete:
            tags: ['Comments']
            summary: Delete comment
            description: Delete a comment (only by the commenter)
            operationId: deleteComment
            responses:
                '200':
                    description: Comment deleted successfully
                '403':
                    description: Not authorized to delete this comment
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                '404':
                    description: Comment not found
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
...